requirements.txt
google_adk
google-cloud-bigquery
google-genai
fastapi
uvicorn
pydantic
python-dotenv
requests